+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New Features & Enhancements
 * Added support for temporary credentials using AWS Session Tokens (See `--s3sessiontoken`.)
+* Experimental: Added support for S3 checksum algorithm selection. (See `--s3checksumalgo`.)
 
 ### General Changes
 * Updated S3 to latest AWS SDK CPP v1.11.580.